moving to implement as a resource action
