+++ app/controllers/api/providers_controller.rb
@@ -21,7 +21,6 @@ module Api
     include Subcollections::Vms
     include Subcollections::Flavors
     include Subcollections::CloudTemplates
-    include Subcollections::ContainerImages
 
     before_action :validate_provider_class
